# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Gamer MineHub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalMods: number;
  totalDownloads: number;
  totalDevelopers: number;
  totalUsers: number;
}

export interface AdminStats {
  totalMods: number;
  totalDownloads: number;
  totalDevelopers: number;
  totalUsers: number;
  pendingMods: number;
  pendingDevelopers: number;
  bannedUsers: number;
}

export type ModCategory = typeof ModCategory[keyof typeof ModCategory];


export const ModCategory = {
  mod: 'mod',
  tool: 'tool',
  texture: 'texture',
  shader: 'shader',
  utility: 'utility',
} as const;

export type ModStatus = typeof ModStatus[keyof typeof ModStatus];


export const ModStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  suspended: 'suspended',
} as const;

export interface Mod {
  id: string;
  title: string;
  description: string;
  category: ModCategory;
  status: ModStatus;
  developerId: string;
  developerName: string;
  /** @nullable */
  developerAvatar?: string | null;
  version: string;
  /** @nullable */
  changelog?: string | null;
  /** @nullable */
  bannerImage?: string | null;
  screenshots?: string[];
  /** @nullable */
  setupVideo?: string | null;
  /** @nullable */
  gameplayVideo?: string | null;
  downloadLink?: string;
  /** @nullable */
  installationGuide?: string | null;
  tags?: string[];
  downloadCount: number;
  likeCount: number;
  featured?: boolean;
  vip?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ModInputCategory = typeof ModInputCategory[keyof typeof ModInputCategory];


export const ModInputCategory = {
  mod: 'mod',
  tool: 'tool',
  texture: 'texture',
  shader: 'shader',
  utility: 'utility',
} as const;

export interface ModInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: ModInputCategory;
  version: string;
  changelog?: string;
  bannerImage?: string;
  screenshots?: string[];
  setupVideo?: string;
  gameplayVideo?: string;
  downloadLink: string;
  installationGuide?: string;
  tags?: string[];
}

export interface ModListResponse {
  mods: Mod[];
  total: number;
}

export type DeveloperStatus = typeof DeveloperStatus[keyof typeof DeveloperStatus];


export const DeveloperStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  banned: 'banned',
} as const;

export type DeveloperBadgesItem = typeof DeveloperBadgesItem[keyof typeof DeveloperBadgesItem];


export const DeveloperBadgesItem = {
  featured: 'featured',
  verified: 'verified',
  vip: 'vip',
} as const;

export interface Developer {
  id: string;
  userId: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  youtubeChannelName?: string | null;
  /** @nullable */
  youtubeLink?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  avatar?: string | null;
  status: DeveloperStatus;
  badges?: DeveloperBadgesItem[];
  followersCount: number;
  uploadsCount: number;
  likesCount: number;
  createdAt: string;
}

export interface DeveloperApplicationInput {
  /** @minLength 1 */
  name: string;
  phone: string;
  youtubeChannelName?: string;
  youtubeLink?: string;
  instagram?: string;
  telegram?: string;
  about: string;
}

export type AppUserRole = typeof AppUserRole[keyof typeof AppUserRole];


export const AppUserRole = {
  user: 'user',
  developer: 'developer',
  admin: 'admin',
} as const;

export type AppUserStatus = typeof AppUserStatus[keyof typeof AppUserStatus];


export const AppUserStatus = {
  active: 'active',
  banned: 'banned',
} as const;

export interface AppUser {
  id: string;
  email: string;
  username: string;
  role: AppUserRole;
  /** @nullable */
  loginProvider?: string | null;
  status: AppUserStatus;
  createdAt: string;
}

export interface Comment {
  id: string;
  modId: string;
  userId: string;
  username: string;
  /** @nullable */
  userAvatar?: string | null;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
  userId: string;
  username: string;
  userAvatar?: string;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export interface FollowResult {
  following: boolean;
  followersCount: number;
}

export interface DownloadTrackInput {
  modId: string;
  /** @nullable */
  userId?: string | null;
}

export interface DownloadTrackResult {
  downloadUrl: string;
  downloadCount: number;
}

export type AnnouncementType = typeof AnnouncementType[keyof typeof AnnouncementType];


export const AnnouncementType = {
  info: 'info',
  warning: 'warning',
  giveaway: 'giveaway',
  update: 'update',
} as const;

export interface Announcement {
  id: string;
  title: string;
  content: string;
  type: AnnouncementType;
  pinned?: boolean;
  createdAt: string;
}

export type AnnouncementInputType = typeof AnnouncementInputType[keyof typeof AnnouncementInputType];


export const AnnouncementInputType = {
  info: 'info',
  warning: 'warning',
  giveaway: 'giveaway',
  update: 'update',
} as const;

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  type: AnnouncementInputType;
  pinned?: boolean;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminSession {
  token: string;
  expiresAt: string;
}

export interface AdminRejectInput {
  reason: string;
}

export type ListModsParams = {
category?: string;
search?: string;
sort?: ListModsSort;
limit?: number;
offset?: number;
};

export type ListModsSort = typeof ListModsSort[keyof typeof ListModsSort];


export const ListModsSort = {
  trending: 'trending',
  latest: 'latest',
  top: 'top',
} as const;

export type ListDevelopersParams = {
sort?: ListDevelopersSort;
};

export type ListDevelopersSort = typeof ListDevelopersSort[keyof typeof ListDevelopersSort];


export const ListDevelopersSort = {
  trending: 'trending',
  top: 'top',
  featured: 'featured',
} as const;

export type AdminListModsParams = {
status?: AdminListModsStatus;
};

export type AdminListModsStatus = typeof AdminListModsStatus[keyof typeof AdminListModsStatus];


export const AdminListModsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  suspended: 'suspended',
} as const;

export type AdminListDevelopersParams = {
status?: AdminListDevelopersStatus;
};

export type AdminListDevelopersStatus = typeof AdminListDevelopersStatus[keyof typeof AdminListDevelopersStatus];


export const AdminListDevelopersStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  banned: 'banned',
} as const;