# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vantage Gamer MineHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /developers:
    get:
      operationId: listDevelopers
      tags: [developers]
      summary: List approved developers
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, top, featured]
      responses:
        "200":
          description: List of developers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Developer"

    post:
      operationId: applyDeveloper
      tags: [developers]
      summary: Submit developer application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Developer"
  /developers/{id}:
    get:
      operationId: getDeveloper
      tags: [developers]
      summary: Get developer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Developer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Developer"
  /developers/{id}/follow:
    post:
      operationId: followDeveloper
      tags: [developers]
      summary: Follow or unfollow a developer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
  /admin/mods:
    get:
      operationId: adminListMods
      tags: [admin]
      summary: Admin - list all mods including pending
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, suspended]
      responses:
        "200":
          description: All mods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModListResponse"
  /admin/mods/{id}/approve:
    post:
      operationId: adminApproveMod
      tags: [admin]
      summary: Admin - approve a mod
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mod approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mod"
  /admin/mods/{id}/reject:
    post:
      operationId: adminRejectMod
      tags: [admin]
      summary: Admin - reject a mod
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRejectInput"
      responses:
        "200":
          description: Mod rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mod"
  /admin/developers:
    get:
      operationId: adminListDevelopers
      tags: [admin]
      summary: Admin - list all developer applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, banned]
      responses:
        "200":
          description: Developer applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Developer"
  /admin/developers/{id}/approve:
    post:
      operationId: adminApproveDeveloper
      tags: [admin]
      summary: Admin - approve developer application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Developer approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Developer"
  /admin/developers/{id}/ban:
    post:
      operationId: adminBanDeveloper
      tags: [admin]
      summary: Admin - ban a developer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Developer banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Developer"
...
  /admin/announcements:
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Admin - create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
...
components:
  schemas:
...
    Mod:
      type: object
      required: [id, title, description, category, status, developerId, developerName, version, downloadCount, likeCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [mod, tool, texture, shader, utility]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        developerId:
          type: string
        developerName:
          type: string
        developerAvatar:
          type: ["string", "null"]
        version:
          type: string
        changelog:
          type: ["string", "null"]
        bannerImage:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
        setupVideo:
          type: ["string", "null"]
        gameplayVideo:
          type: ["string", "null"]
        downloadLink:
          type: string
        installationGuide:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        downloadCount:
          type: integer
        likeCount:
          type: integer
        featured:
          type: boolean
        vip:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ModListResponse:
      type: object
      required: [mods, total]
      properties:
        mods:
          type: array
          items:
            $ref: "#/components/schemas/Mod"
        total:
          type: integer
    Developer:
      type: object
      required: [id, userId, name, status, followersCount, uploadsCount, likesCount, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        youtubeChannelName:
          type: ["string", "null"]
        youtubeLink:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        telegram:
          type: ["string", "null"]
        about:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, banned]
        badges:
          type: array
          items:
            type: string
            enum: [featured, verified, vip]
        followersCount:
          type: integer
        uploadsCount:
          type: integer
        likesCount:
          type: integer
        createdAt:
          type: string
    DeveloperApplicationInput:
      type: object
      required: [name, phone, about]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
        youtubeChannelName:
          type: string
        youtubeLink:
          type: string
        instagram:
          type: string
        telegram:
          type: string
        about:
          type: string
...
    FollowResult:
      type: object
      required: [following, followersCount]
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
...
    Announcement:
      type: object
      required: [id, title, content, type, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [info, warning, giveaway, update]
        pinned:
          type: boolean
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content, type]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum: [info, warning, giveaway, update]
        pinned:
          type: boolean
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminSession:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
        expiresAt:
          type: string
    AdminRejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string