# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Gamer MineHub API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementType } from './announcementType';

export interface Announcement {
  id: string;
  title: string;
  content: string;
  type: AnnouncementType;
  pinned?: boolean;
  createdAt: string;
}