# lib/api-zod/src/generated/types/developer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Gamer MineHub API
 * OpenAPI spec version: 0.1.0
 */
import type { DeveloperBadgesItem } from './developerBadgesItem';
import type { DeveloperStatus } from './developerStatus';

export interface Developer {
  id: string;
  userId: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  youtubeChannelName?: string | null;
  /** @nullable */
  youtubeLink?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  avatar?: string | null;
  status: DeveloperStatus;
  badges?: DeveloperBadgesItem[];
  followersCount: number;
  uploadsCount: number;
  likesCount: number;
  createdAt: string;
}