# lib/api-zod/src/generated/types/mod.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Gamer MineHub API
 * OpenAPI spec version: 0.1.0
 */
import type { ModCategory } from './modCategory';
import type { ModStatus } from './modStatus';

export interface Mod {
  id: string;
  title: string;
  description: string;
  category: ModCategory;
  status: ModStatus;
  developerId: string;
  developerName: string;
  /** @nullable */
  developerAvatar?: string | null;
  version: string;
  /** @nullable */
  changelog?: string | null;
  /** @nullable */
  bannerImage?: string | null;
  screenshots?: string[];
  /** @nullable */
  setupVideo?: string | null;
  /** @nullable */
  gameplayVideo?: string | null;
  downloadLink?: string;
  /** @nullable */
  installationGuide?: string | null;
  tags?: string[];
  downloadCount: number;
  likeCount: number;
  featured?: boolean;
  vip?: boolean;
  createdAt: string;
  updatedAt: string;
}