# lib/api-zod/src/generated/types/modListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Gamer MineHub API
 * OpenAPI spec version: 0.1.0
 */
import type { Mod } from './mod';

export interface ModListResponse {
  mods: Mod[];
  total: number;
}